requirements.txt
spotify-confidence-sdk
openfeature-sdk

